+++ src/components/intersect.tsx
@@ -14,9 +14,9 @@ import processData from '../utils/flattenAndDissolve';
 import { generateColor } from '../utils/genereateColor';
 import generateId from '../utils/generateId';
 import InfoIcon from '@mui/icons-material/Info';
-import { intersectSteps } from '../data/steps/intersectSteps';
+import { intersectSteps } from '../tutorial/steps/intersectSteps';
 import makeStyles from '@mui/styles/makeStyles';
-import Tutorial from './tutorial';
+import Tutorial from '../tutorial/steps/tutorial';
 
 const useStyles = makeStyles({
   hovered: {
@@ -50,8 +50,6 @@ function Intersect(props: {
       const layer2 = selectedLayer2?.geoJSON;
 
       const { processed1, processed2 } = processData(layer1, layer2);
-      console.log('processed1', processed1);
-      console.log('processed2', processed2);
 
       processed1.features.forEach((feature1) => {
         processed2?.features.forEach((feature2) => {
@@ -60,7 +58,6 @@ function Intersect(props: {
             if (feature1.geometry.type === 'Polygon' && feature2.geometry.type === 'Polygon') {
               const intersection = intersect(feature1.geometry, feature2.geometry);
               // Check that there is an intersection at that its not added before
-              console.log('intersection', intersection);
               if (
                 intersection !== null &&
                 finalIntersections.features.every((feat) => !booleanOverlap(intersection, feat))
