+++ src/tutorial/steps/clipSteps.tsx
