+++ src/tutorial/steps/clip_demo.png
