+++ src/tutorial/steps/tutorial.tsx
@@ -1,13 +1,5 @@
-import React, { useState, ChangeEvent, useRef } from 'react';
-import makeStyles from '@mui/styles/makeStyles';
+import React, { useRef } from 'react';
 import Joyride, { CallBackProps, StoreHelpers } from 'react-joyride';
-
-const useStyles = makeStyles({
-  hovered: {
-    backgroundColor: '#f2f2f2',
-    boxShadow: '0 0 5px rgba(0, 0, 0, 0.3)',
-  },
-});
 
 function Tutorial(props: { runTour: boolean; steps: any; setRunTour: (toggle: boolean) => void }) {
   const joyrideHelpers = useRef<StoreHelpers | null>(null);
